Moved site picker adapter to widgets package
