+++ WooCommerce/src/main/kotlin/com/woocommerce/android/widgets/sitepicker/SitePickerAdapter.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.login.adapter
+package com.woocommerce.android.widgets.sitepicker
 
 import android.content.Context
 import android.support.v7.widget.RecyclerView
@@ -9,12 +9,12 @@ import android.view.ViewGroup
 import android.widget.RadioButton
 import android.widget.TextView
 import com.woocommerce.android.R
-import com.woocommerce.android.ui.login.adapter.SiteListAdapter.SiteViewHolder
 import com.woocommerce.android.util.StringUtils
+import com.woocommerce.android.widgets.sitepicker.SitePickerAdapter.SiteViewHolder
 import kotlinx.android.synthetic.main.site_list_item.view.*
 import org.wordpress.android.fluxc.model.SiteModel
 
-class SiteListAdapter(private val context: Context, private val listener: OnSiteClickListener) :
+class SitePickerAdapter(private val context: Context, private val listener: OnSiteClickListener) :
         RecyclerView.Adapter<SiteViewHolder>() {
     var siteList: List<SiteModel> = ArrayList()
         set(value) {
